concealed a lot of output from commands
A design goal of proto (which I haven't written down anywhere, sorry), is that
it should be very non-verbose, and not barf out tons of unimportant build
output for the user. It's very Perl to do that -- CPAN does it, for example --
and all that output has a certain allure of transparency to it. But I still
think we should go for very minimal output. Tell the user the bare minimum.
(That's also why I prefer to print "Perl 6" instead of "Rakudo", even though
we use the latter name liberally in the script itself. Some people don't care
what the implementation is called, but they will know what Perl 6 is.)