[App::Pls] introduced Ecosystem
The Ecosystem role holds information on all projects known to pls. It's
distinct from the ProjectsState role, which holds information on projects
in various installedness on the present system. When asked, the Ecosystem
gives back a hash of information about a given project, sufficient for
the Fetcher et al. to do their job on the project.