[App::Pls] changed build strategy
Instead of moving all dependencies of a project in lockstep with the project
during build-test-install, Yapsi now sees to it that all dependencies
are installed before building the dependent project. Suggested by jnthn++,
and more in line with how CPAN does things. Makes more sense in many ways.

A test started failing, but it's most likely due to the closures cloning bug
in Rakudo. Skipped it for the time being.