[proof-of-concept] write to log instead of /dev/null
When we run commands, write the commands to a logfile pls.log rather than
to /dev/null. This has the advantage of working on Windows, which doesn't
have a /dev/null. If the command succeeds, we remove the log file. The
log file is always written to the pls directory.