[proof-of-concept] fixed .pm file opening
The ufo part of pls was under-tested, and blew up on first use. It should
open a path not relative to the cached project, but relative to the pls
dir. Also decided that a bare &open is too dangerous to have lying around,
and surrounded it with a try {} block.