[App::Pls] make .fetch non-recursive
This took some hacking, and I'm not sure I got it completely right.
But .fetch now only does the project itself, and not its dependencies,
because the dependencies aren't know until the project is fetched.
Other changes had to be made to make sure we didn't end up recursing
forever with mutually dependent projects.