fix page update for IO::Select
The problem was that IO::Select did not have a description line on github, and
the update script used that instead of the description from the META.info file.

This commit now changes the update script to use the description from META.info,
and only falls back to the github repo if META.info has no "description"
field.