Fix UTF-8 encoding mismatch
When the string is read (from Mojo UA), it's not flagged as utf8. To
then encode_utf8($str) (convert a Str to a Buf, basically), and then
write that buf to a :encoding<UTF-8> handle, causes double-encoding.

Instead, the raw string should be decoded from utf8, which turns on the
utf8 flag. Then writing to the :encoding<utf-8> filehandle will do the
right thing.