Fixes #44 Race Condition with Travis Build Info
We"ll simply continue fetching travis info for 24hrs after the last commit to the repo. The 24hr period gives enough time for both Travis status to update for dists whose travis runs forever and for our builder to run the build cycle (which will take forever when we"ll get more dists).