Redesign dist source-url META checker
The current check gives only 1 hour window of opportunity to notice any issues
with the source-url in META file, as the checker does not run on dists without
fresh commits.

Perform the check on all dists regardless whether we have new commits, but do
not do the check if the source-url matches the repo we fetched the META file
from; since this already guarantees the source-url is correct. This will give
us even better reduction of HTTP requests than doing the checks for dists
with fresh commits only and any issues will stay in the build log until fixed.