Fix duplication of modules during db update run
The tag work changed primary key of Dist table and this made
->update_or_create fail to update existing stuff, since we're not
passing it the primary key. So it dupes the modules during build,
and at the end of the build the originals get assumed to be old,
deleted dists and are tossed.

Fix by using meta URL as primary column.