Make Todo controller lookup authors, not full URLs
Required guessing author when it's not specified, and gave us an
excuse to store a Problem when an author's not set.

This also fixes a bug with _builder being cleared if something's
stored there too early: we don't need to declare it anywhere in
particular, autoviv does the work for us.