Fix failure to fetch AppVeyor status / Add 'pending' status
Turns out we can't just use github-username/repo-name as names for
AppVeyor project as they can differ. Haven't found a way to get to
data via REST API, but the badge URL allows "github/" specifier that
does work. So I fixed that problem by parsing the SVG badges.

That gives another problem: status could be "pending", which for
Travis in the past we actually took as-is and never bothered to
refetch. So I now added a unified "pending" status and made the
post-processor re-fetch data for pending statuses during even a cached
rebuild, same as we do for "unknown" statuses.