Pass a `time` query arg when fetching META file
GitHub overly eagerly servers cached version of the META file, so that
even after a commit is done, it serves the old META for some time.

This interferes with testing during dev, and might also cause inconguency
between the META data we have and latest commit data, since we fetch
META off the Web, not the API call.