Defend against potential commit trigger conflict
If a commit to REAPP goes in in the same batch as NEWDB, we'll
be restarting the app over a db it can't handle, leading to a 500 error.
It's might be safe to REAPP when REBUILD is asked for, but I'm
avoiding REAPPING in that case as well, just in case.